stream/stream-generator/requirements.txt
google-cloud-pubsub==0.33.0
pandas
ratelimit==1.4.1